pyproject.toml
[project]
name = "pokemon-tcg-collection-manager-lmm-interface"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.57.0",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "ruff>=0.8.2",
]
